Move vault push secrets task into a separate role
We'd like to keep any of our imperative steps in ansible. We want to avoid
shell as it is just too fragile for most of our use-cases.

The main goal is that this stuff is fully reusable. Otherwise we end up
with fixes landing in one tree (say common) but never in the other repos.

This initial pass implements the following:
- Add common/ansible/ folder with roles and tasks for each imperative thing we
  need to do.
- We'll start off with common/ansible/roles/vault_utils and the task to
  push secrets
- So common/ansible/roles/vault_utils/ will be a role with no main.yaml task
  but instead a bunch of other specific tasks
- So injecting secrets will be something like:
  - name: Run tasks/other.yaml instead of 'main'
    include_role:
      name: vault_utils
      tasks_from: push_secrets
- In industrial edge and other instead of duplicating that code we can
  import specific tasks like done above